Hold mapped dics in memory
