+++ src/Handlers.hs
@@ -1,9 +1,9 @@
 module Handlers
        ( Title
-       , directSearch
-       , getDubs
+       , Dictionary
+       , searchInRaw
+       , searchInMap
        , makeTextMap
-       , mapMaybeTuple
        , mergeWithNum
        , zipWithRaw
        , zipWithMap
@@ -11,82 +11,59 @@ module Handlers
 
 import           Data.Map (Map)
 import           Data.Text (Text)
-import           System.IO (hPrint, stderr)
 import           Path (File, Abs, Path, filename,fromRelFile)
+import           Data.Maybe (isNothing)
 
-import           Prettify (greenCode, boldCode, blueCode, cyanCode, resetCode, putTextFlush)
+import           Prettify (greenCode, boldCode, blueCode, cyanCode, resetCode)
 
 import qualified Data.Map as Map
 import qualified Data.Text as T
-import qualified Data.Text.IO as IO
 
 
 type Title = Text
 
--- | Show duplicates and write to file.
-getDubs :: IO ()
-getDubs = do
-    berzin <- IO.readFile "dics/03-Berzin"
-    putTextFlush "Berzin file is loaded"
-    let dubs = findDups berzin
-    IO.writeFile "dics/dubs" $ T.pack $ show dubs
-    hPrint stderr dubs
+type Dictionary = Map Text Text
 
--- | Make Map from raw file. Merge duplicates to on key without delete
-makeTextMap :: Text -> Map Text Text
+-- | Make Map from raw file. Merge duplicates to on key without delete.
+makeTextMap :: Text -> Dictionary
 makeTextMap
-    = Map.fromAscListWithKey (\_ a1 a2 -> T.concat ["- ", a1, "\n- ", a2])
+    = Map.fromAscListWithKey (\_ a1 a2 -> T.concat [dub, a1, "\n", dub, a2])
     . map ((\(y,x) -> (y, T.drop 1 x))
     . T.span (<'|'))
     . T.lines
-
--- | List tuples of duplicates only from raw file.
-findDups :: Text -> [(Text,Text)]
-findDups
-    = dups
-    . map ((\(y,x) -> (y, T.drop 1 x))
-    . T.span (<'|'))
-    . T.lines
-
--- | Get duplicate tuples only.
-dups :: Eq k => [(k,v)] -> [(k,v)]
-dups []     = []
-dups (x:xs) =  d ++ dups r
   where
-    (d,r) = (resultDub x, filter (\(y,_) -> fst x /= y) xs)
-    isDub = filter (\(y,_) -> fst x == y) xs
-    resultDub first | null isDub = []
-                    | otherwise  = first : isDub
-
--- | Remove empty answers
-mapMaybeTuple :: (a -> Maybe b) -> [(a,t)] -> [(b,t)]
-mapMaybeTuple _ [] = []
-mapMaybeTuple f ((x, t):xs) =
-    let rs = mapMaybeTuple f xs in
-    case (f x, t) of
-        (Nothing, _) -> rs
-        (Just r, n)  -> (r,n):rs
+    dub = (cyanCode <> "༔ " <> resetCode)
 
 -- | Combine dictionary titles with mapped dictionaries.
-zipWithMap :: [Text] -> [Path Abs File] -> [(Map Text Text, Title)]
+zipWithMap :: [Text] -> [Path Abs File] -> [(Dictionary, Title)]
 zipWithMap texts files = zip mapped titles
   where
-    mapped :: [Map Text Text]
+    mapped :: [Dictionary]
     mapped = map makeTextMap texts
 
-    titles :: [Text]
+    titles :: [Title]
     titles = map (T.drop 3 . T.pack . fromRelFile . filename) files
+
+-- Search in mapped dictionary.
+searchInMap :: Text -> [(Dictionary, Title)] -> [(Text, Title)]
+searchInMap query mapped = [(text, title) | (Just text, title) <- searched]
+  where
+    searched :: [(Maybe Text, Title)]
+    searched = foldl (\ acc (x,y) -> if isNothing (search x) then acc else (search x, y) : acc) [] mapped
+
+    search :: Dictionary -> Maybe Text
+    search = Map.lookup query
 
 -- | Combine answers with numbering for raw text.
 zipWithRaw :: [Text] -> [Path Abs File] -> [(Text, Title)]
 zipWithRaw texts files = zip texts titles
   where
     titles :: [Text]
-    titles = map ((\x -> blueCode <> boldCode<> x <> resetCode) . T.drop 3 . T.pack . fromRelFile . filename) files
+    titles = map (T.drop 3 . T.pack . fromRelFile . filename) files
 
 -- | Search in raw dictionary files.
-directSearch :: Text -> [(Text, Title)] -> [(Text, Title)]
-directSearch query = foldl (\ acc (x,y) -> if search x == "" then acc else (search x, y) : acc) []
+searchInRaw :: Text -> [(Text, Title)] -> [(Text, Title)]
+searchInRaw query = foldl (\ acc (x,y) -> if search x == "" then acc else (search x, y) : acc) []
   where
     search :: Text -> Text
     search
@@ -95,16 +72,20 @@ directSearch query = foldl (\ acc (x,y) -> if search x == "" then acc else (sear
         . filter (T.isPrefixOf (T.append query "|"))
         . T.lines
 
-
-mergeWithNum :: [(Text, Text)] -> Text
+-- | Add numbers and flatten.
+mergeWithNum :: [(Text, Title)] -> Text
 mergeWithNum = T.unlines . zipWith flatten numbers
   where
     -- uncomment to use mapped data
     -- ascValues :: [(Text, Text)]
     -- ascValues = map (\(v,t) -> (T.unlines . reverse . T.lines $ v, t)) dscValues
 
+    -- Add numbers.
     numbers :: [Text]
     numbers = map ((\x -> greenCode <> T.append (T.pack x) ". " <> resetCode) . show) [1::Int ..]
 
-    flatten :: Text -> (Text, Text) -> Text
-    flatten n (v, t) = T.append (T.append n (T.append t "\n")) v
+    flatten :: Text -> (Text, Title) -> Text
+    flatten n (v, t) = T.append (T.append n (T.append (prettyT t) "\n")) v
+    -- Paint title
+    prettyT :: Title -> Title
+    prettyT t = blueCode <> boldCode <> t <> resetCode
